+++ NEWS.md
@@ -4,6 +4,9 @@
 
 ## New features
 
+* Tree fits are now also accepted from  from the `partykit` package
+(`constparty` objects), instead of only from the `rpart` package.
+
 * Functions `fusedTree` and `PenOpt` now take a new optional input argument
 `symFusion`. Setting this argument to `FALSE` allows for asymmetric fusion,
 where omics effects in  nodes having similar predictions of the response
@@ -11,7 +14,7 @@ where omics effects in  nodes having similar predictions of the response
 to omics effects in nodes having more distinct predictions. 
 
 * For memory efficiency, Matrix algebra is now conduced using sparse matrices
-from the [`Matrix`] package when `X` is high-dimensional.
+from the `Matrix` package when `X` is high-dimensional.
 
 # fusedTree 1.0.1